Create gradle.yml
Add CI with GitHub action 

This workflow will build a java 1.13 project with gradle on a system running the latest version of ubuntu. All tests are run and must pass during the build process

This workflow is triggered on a push or PR to any branch. 

NO BAD CODE ALLOWED 🙅🏼‍♂️🙅🏼‍♂️🙅🏼‍♂️